verify_submission.py: delay git clone
Cloning the VK-GL-CTS repository can take a long time, so defer
doing that until at least some of the sanity checks have been done
(e.g. that we know the tar file is legal). This prevents the user
from having to wait 5 minutes (or longer!) to discover that they've
mistyped the file name.